+++ src/plugins/data/common/es_query/filters/build_filter.test.ts
@@ -18,7 +18,7 @@
  */
 
 import { buildFilter, FilterStateStore, FILTERS } from '.';
-import { stubIndexPattern, stubFields } from '../../../public/stubs';
+import { stubIndexPattern, stubFields } from '../../../common/stubs';
 
 describe('buildFilter', () => {
   it('should build phrase filters', () => {
