+++ src/plugins/data/common/es_query/filters/get_index_pattern_from_filter.test.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { stubIndexPattern, phraseFilter } from 'src/plugins/data/public/stubs';
+import { stubIndexPattern, phraseFilter } from 'src/plugins/data/common/stubs';
 import { getIndexPatternFromFilter } from './get_index_pattern_from_filter';
 
 describe('getIndexPatternFromFilter', () => {
