+++ src/plugins/data/common/es_query/kuery/node_types/named_arg.ts
@@ -21,7 +21,7 @@ import _ from 'lodash';
 import * as ast from '../ast';
 import { nodeTypes } from '../node_types';
 import { NamedArgTypeBuildNode } from './types';
-import { JsonObject } from '../../../../../kibana_utils/public';
+import { JsonObject } from '../../../../../kibana_utils/common';
 
 export function buildNode(name: string, value: any): NamedArgTypeBuildNode {
   const argumentNode =
