+++ src/plugins/data/common/index_patterns/field.stub.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { IFieldType } from '../../../../plugins/data/public';
+import { IFieldType } from '.';
 
 export const stubFields: IFieldType[] = [
   {
