+++ src/plugins/expressions/common/execution/execution.test.ts
@@ -20,7 +20,7 @@
 import { Execution } from './execution';
 import { parseExpression, ExpressionAstExpression } from '../ast';
 import { createUnitTestExecutor } from '../test_helpers';
-import { ExpressionFunctionDefinition } from '../../public';
+import { ExpressionFunctionDefinition } from '../../common';
 import { ExecutionContract } from './execution_contract';
 
 beforeAll(() => {
