+++ x-pack/plugins/security/common/licensing/license_service.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { of, BehaviorSubject } from 'rxjs';
-import { licensingMock } from '../../../licensing/public/mocks';
+import { licenseMock } from '../../../licensing/common/licensing.mock';
 import { SecurityLicenseService } from './license_service';
 
 describe('license features', function () {
@@ -29,7 +29,7 @@ describe('license features', function () {
   });
 
   it('should display error when X-Pack is unavailable', () => {
-    const rawLicenseMock = licensingMock.createLicenseMock();
+    const rawLicenseMock = licenseMock.createLicenseMock();
     rawLicenseMock.isAvailable = false;
     const serviceSetup = new SecurityLicenseService().setup({
       license$: of(rawLicenseMock),
@@ -50,7 +50,7 @@ describe('license features', function () {
   });
 
   it('should notify consumers of licensed feature changes', () => {
-    const rawLicenseMock = licensingMock.createLicenseMock();
+    const rawLicenseMock = licenseMock.createLicenseMock();
     rawLicenseMock.isAvailable = false;
     const rawLicense$ = new BehaviorSubject(rawLicenseMock);
     const serviceSetup = new SecurityLicenseService().setup({
@@ -79,7 +79,7 @@ describe('license features', function () {
         ]
       `);
 
-      rawLicense$.next(licensingMock.createLicenseMock());
+      rawLicense$.next(licenseMock.createLicenseMock());
       expect(subscriptionHandler).toHaveBeenCalledTimes(2);
       expect(subscriptionHandler.mock.calls[1]).toMatchInlineSnapshot(`
         Array [
@@ -103,7 +103,7 @@ describe('license features', function () {
   });
 
   it('should show login page and other security elements, allow RBAC but forbid paid features if license is basic.', () => {
-    const mockRawLicense = licensingMock.createLicense({
+    const mockRawLicense = licenseMock.createLicense({
       features: { security: { isEnabled: true, isAvailable: true } },
     });
 
@@ -129,7 +129,7 @@ describe('license features', function () {
   });
 
   it('should not show login page or other security elements if security is disabled in Elasticsearch.', () => {
-    const mockRawLicense = licensingMock.createLicense({
+    const mockRawLicense = licenseMock.createLicense({
       features: { security: { isEnabled: false, isAvailable: true } },
     });
 
@@ -151,7 +151,7 @@ describe('license features', function () {
   });
 
   it('should allow role mappings, access agreement and sub-feature privileges, but not DLS/FLS if license = gold', () => {
-    const mockRawLicense = licensingMock.createLicense({
+    const mockRawLicense = licenseMock.createLicense({
       license: { mode: 'gold', type: 'gold' },
       features: { security: { isEnabled: true, isAvailable: true } },
     });
@@ -174,7 +174,7 @@ describe('license features', function () {
   });
 
   it('should allow to login, allow RBAC, role mappings, access agreement, sub-feature privileges, and DLS if license >= platinum', () => {
-    const mockRawLicense = licensingMock.createLicense({
+    const mockRawLicense = licenseMock.createLicense({
       license: { mode: 'platinum', type: 'platinum' },
       features: { security: { isEnabled: true, isAvailable: true } },
     });
@@ -197,7 +197,7 @@ describe('license features', function () {
   });
 
   it('should allow all basic features + audit logging for standard license', () => {
-    const mockRawLicense = licensingMock.createLicense({
+    const mockRawLicense = licenseMock.createLicense({
       license: { mode: 'standard', type: 'standard' },
       features: { security: { isEnabled: true, isAvailable: true } },
     });
